% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_surfaceome}
\alias{annotate_surfaceome}
\title{Title}
\usage{
annotate_surfaceome(
  x,
  genes,
  surfaceome = surfaceome::surfy,
  surfaceome_identifier = `UniProt gene`
)
}
\arguments{
\item{x}{A matrix of gene expression}

\item{genes}{The list of gene names (possibly separated by \verb{;})}

\item{surfaceome}{The surfaceome table to join into results}

\item{surfaceome_identifier}{The column name to use as identifier. Should match the type of \code{genes}.}
}
\description{
Title
}
