% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calculate_gene_statistics}
\alias{calculate_gene_statistics}
\title{Calculate gene statistics for a matrix}
\usage{
calculate_gene_statistics(x)
}
\arguments{
\item{x}{A matrix to calculate gene statistics}
}
\description{
Calculate gene statistics for a matrix
}
