% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overexpressed_surfaceome.R
\name{overexpressed_surfaceome}
\alias{overexpressed_surfaceome}
\title{Calculate the over-expressed surfaceome}
\usage{
overexpressed_surfaceome(x, which = c("surfy", "cspa"))
}
\arguments{
\item{x}{A matrix of expression data}

\item{which}{Which surfaceome definition to use (surfy or cspa)}
}
\value{
A ordering of the gene expression by over-expression (median expression).
}
\description{
Calculate the over-expressed surfaceome
}
