% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_surfaceome}
\alias{subset_surfaceome}
\title{Subset}
\usage{
subset_surfaceome(x, surfaceome = surfaceome::surfy$`UniProt gene`)
}
\arguments{
\item{surfaceome}{}
}
\description{
The surfaceome table (UniProt gene) has only one gene identifier per line (although
it has multiple lines with the same gene symbol). The experimental data, however,
may have gene smbols that are actually a list (separated with ;). So subsetting to
the surfaceome is a little more complicated.
}
